product
 Conflicts:
	app/target/app-0.0.1-SNAPSHOT.jar
	app/target/app-0.0.1-SNAPSHOT.jar.original
	app/target/classes/com/example/multi/app/controller/ProductController.class
	app/target/classes/com/example/multi/app/controller/TestController.class
	app/target/classes/com/example/multi/app/controller/UploadController.class
	app/target/classes/com/example/multi/app/controller/test.class
	console/target/console-0.0.1-SNAPSHOT.jar
	console/target/console-0.0.1-SNAPSHOT.jar.original
	module/target/module-0.0.1-SNAPSHOT.jar

+++ console/src/main/java/com/example/multi/console/controller/ProductController.java
@@ -1,12 +1,18 @@
 package com.example.multi.console.controller;
 
 import com.example.multi.console.annotation.RequireLogin;
-import com.example.multi.console.domain.*;
+import com.example.multi.console.domain.category.ConsoleCreateVO;
+import com.example.multi.console.domain.category.ConsoleUpdateVO;
+import com.example.multi.console.domain.product.ProductConCellVO;
+import com.example.multi.console.domain.product.ProductConInfoVO;
+import com.example.multi.console.domain.product.ProductConListVO;
 import com.example.multi.module.category.entity.Category;
 import com.example.multi.module.category.service.CategoryService;
 import com.example.multi.module.product.entity.Product;
 import com.example.multi.module.product.service.ProductService;
-import com.example.multi.module.response.entity.Response;
+import com.example.multi.module.user.entity.User;
+import com.example.multi.module.utils.BaseUtils;
+import com.example.multi.module.utils.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,123 +23,6 @@ import java.util.Arrays;
 import java.util.List;
 
 
-//@RestController
-//public class ProductController {
-//    @Autowired
-//    private ProductService service;
-//    @Autowired
-//    private CategoryService categoryservice;
-//
-//    @RequireLogin
-//    @RequestMapping("/product/list")
-//    public ProductConListVO page(@RequestParam("page") Integer page,
-//                                 @RequestParam(value = "pageSize",defaultValue = "5")Integer pageSize,
-//                                 @RequestParam(value = "keyword",defaultValue = "" ) String keyword){
-//
-//
-//        List<Product> products = service.getPage(page   ,pageSize,keyword);
-//
-//        List<ProductConCellVO>  productConOnlyVOS = new ArrayList<>();
-//        for(Product product : products){
-//
-//            ProductConCellVO productConOnlyVO = new ProductConCellVO();
-//            productConOnlyVO.setId(product.getId());
-//            productConOnlyVO.setTitle(product.getTitle());
-//            String[]  image = product.getImages().split("\\$");
-//            productConOnlyVO.setImage(image[0]);
-//            productConOnlyVO.setInfo(product.getInfo());
-//            productConOnlyVO.setPrice(product.getPrice());
-//            productConOnlyVOS.add(productConOnlyVO);
-//        }
-//        ProductConListVO productConListVO = new ProductConListVO();
-//        productConListVO.setTotal(service.pageCount(keyword));
-//        productConListVO.setPageSize(pageSize);
-//        productConListVO.setList(productConOnlyVOS);
-//        return productConListVO;
-//    }
-//
-//
-//    @RequestMapping("/product/info")
-//    public ProductConInfoVO getInfo(@RequestParam(name = "id") BigInteger id) {
-//        SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-//        ProductConInfoVO productConInfoVO = new ProductConInfoVO();
-//
-//
-//                Product product = service.getById(id);
-//        if (product == null) {
-//            productConInfoVO.setTitle("未找到对应的产品信息");
-//            return productConInfoVO;
-//        }
-//        Category category = categoryservice.getById(product.getCategoryId());
-//        if (category == null) {
-//            productConInfoVO.setTitle("未找到对应的分类信息");
-//            return productConInfoVO;
-//        }
-//                productConInfoVO.setTitle(product.getTitle());
-//                String[] image = product.getImages().split("\\$");
-//                List<String> imageList = Arrays.asList(image);
-//                productConInfoVO.setImages(imageList);
-//                productConInfoVO.setName(product.getName());
-//                productConInfoVO.setInfo(product.getInfo());
-//                productConInfoVO.setPrice(product.getPrice());
-//                productConInfoVO.setDetailedTitle(product.getDetailedTitle());
-//                productConInfoVO.setDetailed(product.getDetailed());
-//                productConInfoVO.setCreateTime(dateFormat.format(product.getCreateTime() * 1000l));
-//                productConInfoVO.setUpdateTime(dateFormat.format(product.getUpdateTime() * 1000l));
-//                productConInfoVO.setCategoryId(product.getCategoryId());
-//        return productConInfoVO;
-//    }
-//
-//    @RequestMapping("/product/create")
-//    public ConsoleCreateVO productCreate(@RequestParam(name = "name") String name,
-//                                                         @RequestParam(name = "title") String title,
-//                                                         @RequestParam(name = "images") String images,
-//                                                         @RequestParam(name = "info") String info,
-//                                                         @RequestParam(name = "price") Integer price,
-//                                                         @RequestParam(name = "detailedTitle") String detailedTitle,
-//                                                         @RequestParam(name = "detailed") String detailed,
-//                                                         @RequestParam(name = "categoryId")BigInteger categoryId) {
-//        ConsoleCreateVO consoleCreateVO = new ConsoleCreateVO();
-//        try {
-//            BigInteger result = service.edit(null, name, title, images, info, price, detailedTitle, detailed,categoryId);
-//            consoleCreateVO.setTips(result != null ? "成功" : "失败");
-//            consoleCreateVO.setId(result);
-//        } catch (Exception exception) {
-//
-//            consoleCreateVO.setTips(exception.getMessage());
-//        }
-//        return consoleCreateVO;
-//    }
-//
-//    @RequestMapping("/product/update")
-//    public ConsoleUpdateVO Update(@RequestParam(name = "id") BigInteger id,
-//                                  @RequestParam(name = "name") String name,
-//                                  @RequestParam(name = "title") String title,
-//                                  @RequestParam(name = "images") String images,
-//                                  @RequestParam(name = "info") String info,
-//                                  @RequestParam(name = "price") Integer price,
-//                                  @RequestParam(name = "detailedTitle") String detailedTitle,
-//                                  @RequestParam(name = "detailed") String detailed,
-//                                  @RequestParam(name = "categoryId") BigInteger categoryId){
-//        ConsoleUpdateVO consoleUpdateVO = new ConsoleUpdateVO();
-//        try {
-//            BigInteger result = service.edit(id, name, title, images, info, price, detailedTitle, detailed,categoryId);
-//            consoleUpdateVO.setTips(result != null ? "成功" : "失败");
-//            consoleUpdateVO.setId(id);
-//        } catch (Exception exception) {
-//            consoleUpdateVO.setTips(exception.getMessage());
-//        }
-//        return consoleUpdateVO;
-//    }
-//
-//    @RequestMapping("/product/delete")
-//    public ConsoleDeleteVO Deleted(@RequestParam(name = "id") BigInteger id){
-//        int result = service.delete(id);
-//        ConsoleDeleteVO consoleDeleteVO = new ConsoleDeleteVO();
-//        consoleDeleteVO.setTips(result  == 1  ? "成功" : "失败");
-//        return  consoleDeleteVO;
-//    }
-
 @RestController
 public class ProductController {
     @Autowired
@@ -141,11 +30,14 @@ public class ProductController {
     @Autowired
     private CategoryService categoryService;
 
-    @RequireLogin
     @RequestMapping("/product/list")
-    public Response<ProductConListVO> page(@RequestParam("page") Integer page,
-                                           @RequestParam(value = "pageSize", defaultValue = "5") Integer pageSize,
-                                           @RequestParam(value = "keyword", defaultValue = "") String keyword) {
+    public Response page(@RequireLogin User loginUser,
+                         @RequestParam("page") Integer page,
+                         @RequestParam(value = "pageSize", defaultValue = "5") Integer pageSize,
+                         @RequestParam(value = "keyword", defaultValue = "") String keyword) {
+        if (BaseUtils.isEmpty(loginUser)) {
+            return new Response(1002);
+        }
         List<Product> products = service.getPage(page, pageSize, keyword);
 
         List<ProductConCellVO> productConOnlyVOS = new ArrayList<>();
@@ -165,19 +57,23 @@ public class ProductController {
         productConListVO.setPageSize(pageSize);
         productConListVO.setList(productConOnlyVOS);
 
-        return Response.success(productConListVO);
+        return new Response(1001, productConListVO);
     }
 
     @RequestMapping("/product/info")
-    public Response<ProductConInfoVO> getInfo(@RequestParam(name = "id") BigInteger id) {
+    public Response getInfo(@RequireLogin User loginUser,
+                            @RequestParam(name = "id") BigInteger id) {
+        if (BaseUtils.isEmpty(loginUser)) {
+            return new Response(1002);
+        }
         Product product = service.getById(id);
         if (product == null) {
-            return Response.error(2001, "未找到对应的产品信息");
+            return new Response(3052);
         }
 
         Category category = categoryService.getById(product.getCategoryId());
         if (category == null) {
-            return Response.error(2002, "未找到对应的分类信息");
+            return new Response(3053);
         }
 
         ProductConInfoVO productConInfoVO = new ProductConInfoVO();
@@ -193,69 +89,77 @@ public class ProductController {
         productConInfoVO.setUpdateTime(new SimpleDateFormat("yyyy-MM-dd HH:mm:ss").format(product.getUpdateTime()));
         productConInfoVO.setCategoryId(product.getCategoryId());
 
-        return Response.success(productConInfoVO);
+        return new Response(1001, productConInfoVO);
     }
 
     @RequestMapping("/product/create")
-    public Response<ConsoleCreateVO> productCreate(@RequestParam(name = "name") String name,
-                                                   @RequestParam(name = "title") String title,
-                                                   @RequestParam(name = "images") String images,
-                                                   @RequestParam(name = "info") String info,
-                                                   @RequestParam(name = "price") Integer price,
-                                                   @RequestParam(name = "detailedTitle") String detailedTitle,
-                                                   @RequestParam(name = "detailed") String detailed,
-                                                   @RequestParam(name = "categoryId") BigInteger categoryId) {
+    public Response productCreate(@RequireLogin User loginUser,
+                                  @RequestParam(name = "name") String name,
+                                  @RequestParam(name = "title") String title,
+                                  @RequestParam(name = "images") String images,
+                                  @RequestParam(name = "info") String info,
+                                  @RequestParam(name = "price") Integer price,
+                                  @RequestParam(name = "detailedTitle") String detailedTitle,
+                                  @RequestParam(name = "detailed") String detailed,
+                                  @RequestParam(name = "categoryId") BigInteger categoryId) {
+        if (BaseUtils.isEmpty(loginUser)) {
+            return new Response(1002);
+        }
         ConsoleCreateVO consoleCreateVO = new ConsoleCreateVO();
         try {
             BigInteger result = service.edit(null, name, title, images, info, price, detailedTitle, detailed, categoryId);
             if (result != null) {
                 consoleCreateVO.setId(result);
-                return Response.success(consoleCreateVO);
+                return new Response(1001, consoleCreateVO);
             } else {
-                return Response.error(4004, "网络繁忙");
+                return new Response(3001);
             }
         } catch (Exception exception) {
-            return Response.error(3001, "产品新增失败:" + exception.getMessage());
+            return new Response(3051, exception.getMessage());
         }
     }
 
     @RequestMapping("/product/update")
-    public Response<ConsoleUpdateVO> update(@RequestParam(name = "id") BigInteger id,
-                                            @RequestParam(name = "name") String name,
-                                            @RequestParam(name = "title") String title,
-                                            @RequestParam(name = "images") String images,
-                                            @RequestParam(name = "info") String info,
-                                            @RequestParam(name = "price") Integer price,
-                                            @RequestParam(name = "detailedTitle") String detailedTitle,
-                                            @RequestParam(name = "detailed") String detailed,
-                                            @RequestParam(name = "categoryId") BigInteger categoryId) {
+    public Response update(@RequireLogin User loginUser,
+                           @RequestParam(name = "id") BigInteger id,
+                           @RequestParam(name = "name") String name,
+                           @RequestParam(name = "title") String title,
+                           @RequestParam(name = "images") String images,
+                           @RequestParam(name = "info") String info,
+                           @RequestParam(name = "price") Integer price,
+                           @RequestParam(name = "detailedTitle") String detailedTitle,
+                           @RequestParam(name = "detailed") String detailed,
+                           @RequestParam(name = "categoryId") BigInteger categoryId) {
+        if (BaseUtils.isEmpty(loginUser)) {
+            return new Response(1002);
+        }
         ConsoleUpdateVO consoleUpdateVO = new ConsoleUpdateVO();
         try {
             BigInteger result = service.edit(id, name, title, images, info, price, detailedTitle, detailed, categoryId);
             if (result != null) {
                 consoleUpdateVO.setId(id);
-                return Response.success(consoleUpdateVO);
+                return new Response(1001, consoleUpdateVO);
             } else {
-                return Response.error(4004, "网络繁忙");
+                return new Response(3002);
             }
         } catch (Exception exception) {
-            return Response.error(3002, "产品更新失败:" + exception.getMessage());
+            return new Response(3051, exception.getMessage());
         }
     }
 
     @RequestMapping("/product/delete")
-    public Response<ConsoleDeleteVO> deleted(@RequestParam(name = "id") BigInteger id) {
+    public Response deleted(@RequireLogin User loginUser,
+                            @RequestParam(name = "id") BigInteger id) {
+        if (BaseUtils.isEmpty(loginUser)) {
+            return new Response(1002);
+        }
         int result = service.delete(id);
         if (result == 1) {
-            return Response.success(new ConsoleDeleteVO());
+            return new Response(1001);
         } else {
-            return Response.error(3003, "产品删除失败");
+            return new Response(3052);
         }
     }
-
-
-
-
 
 
 }
